+++ android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/StackHeaderCoordinatorLayout.kt
@@ -2,14 +2,21 @@ package com.swmansion.rnscreens.gamma.stack.header
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.view.View
 import android.view.ViewGroup.LayoutParams.MATCH_PARENT
 import android.widget.FrameLayout
 import androidx.activity.OnBackPressedDispatcherOwner
+import androidx.appcompat.view.ContextThemeWrapper
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import com.facebook.react.bridge.ReactContext
-import com.swmansion.rnscreens.gamma.stack.header.config.OnHeaderConfigAttachListener
-import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderConfigDelegate
-import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderConfigProviding
+import com.google.android.material.R
+import com.google.android.material.appbar.AppBarLayout
+import com.swmansion.rnscreens.gamma.stack.header.config.OnHeaderConfigurationAttachListener
+import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderConfigurationObserver
+import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderConfigurationProviding
+import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderDelegate
+import com.swmansion.rnscreens.gamma.stack.header.config.StackHeaderUpdateFlags
+import com.swmansion.rnscreens.gamma.stack.header.subview.StackHeaderSubviewProviding
 import com.swmansion.rnscreens.gamma.stack.header.toolbar.StackHeaderToolbarMenuItemOptions
 import com.swmansion.rnscreens.gamma.stack.screen.StackScreen
 import java.lang.ref.WeakReference
@@ -18,114 +25,279 @@ import java.lang.ref.WeakReference
 internal class StackHeaderCoordinatorLayout(
     context: Context,
     internal val stackScreen: StackScreen,
-    canNavigateBack: Boolean,
+    private val canNavigateBack: Boolean,
 ) : CoordinatorLayout(context) {
-    private val headerCoordinator =
-        StackHeaderCoordinator(
-            context = context,
-            canNavigateBack = canNavigateBack,
-            onHeaderHeightChanged = { headerHeight ->
-                stackScreen.updateStateIfNeeded(y = headerHeight)
-            },
-            onNavigationIconClick = {
-                val activity =
-                    (stackScreen.context as? ReactContext)?.currentActivity
-                        as? OnBackPressedDispatcherOwner
-                activity?.onBackPressedDispatcher?.onBackPressed()
-            },
+    private val wrappedContext =
+        ContextThemeWrapper(
+            context,
+            R.style.Theme_Material3_DayNight_NoActionBar,
         )
 
-    /**
-     * This callback is used to detect when header config is attached.
-     * This allows us to configure the delegate for header config interactions.
-     */
-    private val onHeaderConfigAttach =
-        OnHeaderConfigAttachListener { config ->
-            handleHeaderConfigAttach(config)
-        }
+    private val applicator = StackHeaderApplicator(wrappedContext)
 
-    private var isHeaderUpdatePending = false
+    private var currentProvider: StackHeaderConfigurationProviding? = null
+    private var currentDelegate: StackHeaderDelegate? = null
+    private var appBarLayout: StackHeaderAppBarLayout? = null
 
-    // Read currentConfig when the runnable executes, not when it's posted,
-    // to avoid applying a stale config that was swapped out in the meantime.
-    private val headerUpdateRunnable =
-        Runnable {
-            isHeaderUpdatePending = false
-            headerCoordinator.applyHeaderConfig(this, currentConfig)
-        }
+    private var toolbarMenuForwardIdMap = emptyMap<String, Int>()
+    private var toolbarMenuReverseIdMap = emptyMap<Int, String>()
 
-    /**
-     * Single delegate that owns all interactions flowing from [StackHeaderConfig] to this layout.
-     * [onConfigChange] is batched via [post] to coalesce rapid updates.
-     * [onMenuItemUpdate] is dispatched immediately — commands must not be deferred.
-     */
-    private val headerConfigDelegate =
-        object : StackHeaderConfigDelegate {
-            override fun onConfigChange(config: StackHeaderConfigProviding) {
-                if (!isHeaderUpdatePending) {
-                    isHeaderUpdatePending = true
-                    post(headerUpdateRunnable)
-                }
-            }
+    private val onNavigationIconClick: () -> Unit = {
+        val activity =
+            (stackScreen.context as? ReactContext)?.currentActivity
+                as? OnBackPressedDispatcherOwner
+        activity?.onBackPressedDispatcher?.onBackPressed()
+    }
+
+    // region Configuration observer
+
+    private val configObserver =
+        object : StackHeaderConfigurationObserver {
+            override fun onConfigChanged(
+                config: StackHeaderConfigurationProviding,
+                flags: StackHeaderUpdateFlags,
+            ) = processUpdate(config, flags)
 
             override fun onMenuItemUpdate(
                 id: String,
                 options: StackHeaderToolbarMenuItemOptions,
             ) {
-                headerCoordinator.handleMenuItemUpdate(id, options)
+                val toolbar = appBarLayout?.toolbar ?: return
+                applicator.updateToolbarMenuItem(toolbar, toolbarMenuForwardIdMap, id, options)
             }
         }
 
-    private var currentConfig: StackHeaderConfigProviding? = null
+    // endregion
+
+    // region Config attach / detach
+
+    private val onHeaderConfigAttach =
+        OnHeaderConfigurationAttachListener { provider, delegate ->
+            handleHeaderConfigAttach(provider, delegate)
+        }
+
+    private fun handleHeaderConfigAttach(
+        provider: StackHeaderConfigurationProviding?,
+        delegate: StackHeaderDelegate?,
+    ) {
+        currentProvider?.setConfigObserver(null)
+
+        currentProvider = provider
+        currentDelegate = delegate
+
+        provider?.setConfigObserver(configObserver)
+
+        if (provider != null) {
+            processUpdate(provider, StackHeaderUpdateFlags.ALL)
+        } else {
+            removeHeader()
+        }
+    }
+
+    // endregion
+
+    // region Init
 
     internal var stackScreenWrapper: FrameLayout
 
     init {
-        // Needed when Transition API is in use to ensure that shadows do not disappear,
-        // views do not jump around the screen and whole subtree is animated as a whole.
         isTransitionGroup = true
 
-        // Due to how we're synchronizing native & Yoga layout (via contentOriginOffset on
-        // StackScreen), we can't use StackScreen directly as a child of CoordinatorLayout because
-        // SurfaceMountingManager will override Y offset (that depends on the header height) with
-        // Y=0. If we wrap StackScreen in another view, as Y is relative to parent view, value set
-        // by Yoga will be correct.
         stackScreenWrapper = FrameLayout(context).apply { addView(stackScreen) }
         addView(
             stackScreenWrapper,
             LayoutParams(MATCH_PARENT, MATCH_PARENT),
         )
 
-        stackScreen.onHeaderConfigAttachListener = WeakReference(onHeaderConfigAttach)
-        handleHeaderConfigAttach(stackScreen.headerConfig)
+        stackScreen.onHeaderConfigurationAttachListener = WeakReference(onHeaderConfigAttach)
+        handleHeaderConfigAttach(stackScreen.headerConfig, stackScreen.headerConfig)
     }
 
-    private fun handleHeaderConfigAttach(config: StackHeaderConfigProviding?) {
-        // Disconnect old config to prevent spurious updates from a detached config
-        currentConfig?.removeDelegate(headerConfigDelegate)
-        currentConfig = config
+    // endregion
 
-        config?.setDelegate(headerConfigDelegate)
+    // region Flag-gated dispatch
 
-        // We run this even if config is null to properly remove the header if config
-        // is removed in runtime.
-        headerCoordinator.applyHeaderConfig(this, config)
+    private fun processUpdate(
+        provider: StackHeaderConfigurationProviding,
+        flags: StackHeaderUpdateFlags,
+    ) {
+        if (flags.needsRebuild) {
+            resetHeader()
+            if (!provider.hidden) {
+                val appBar = applicator.rebuild(this, provider, canNavigateBack, onNavigationIconClick)
+                appBarLayout = appBar
+                attachAppBarListeners(appBar)
+
+                applicator.applyTitle(appBar, provider)
+                applicator.applyBackButton(appBar.toolbar, provider, canNavigateBack, onNavigationIconClick)
+                applicator.applyScrollFlags(appBar, provider)
+
+                val (fwd, rev) = applicator.rebuildToolbarMenu(
+                    appBar.toolbar,
+                    provider.toolbarMenuItems,
+                ) { id -> currentDelegate?.onMenuItemClick(id) }
+                toolbarMenuForwardIdMap = fwd
+                toolbarMenuReverseIdMap = rev
+            } else {
+                removeContentBehavior()
+                requestLayout()
+            }
+            syncShadowState()
+            return
+        }
+
+        val appBar = appBarLayout ?: return
+
+        if (flags.containsAny(StackHeaderUpdateFlags.TITLE))
+            applicator.applyTitle(appBar, provider)
+        if (flags.containsAny(StackHeaderUpdateFlags.BACK_BUTTON))
+            applicator.applyBackButton(appBar.toolbar, provider, canNavigateBack, onNavigationIconClick)
+        if (flags.containsAny(StackHeaderUpdateFlags.SCROLL_FLAGS))
+            applicator.applyScrollFlags(appBar, provider)
+        if (flags.containsAny(StackHeaderUpdateFlags.TOOLBAR_MENU)) {
+            val (fwd, rev) = applicator.rebuildToolbarMenu(
+                appBar.toolbar,
+                provider.toolbarMenuItems,
+            ) { id -> currentDelegate?.onMenuItemClick(id) }
+            toolbarMenuForwardIdMap = fwd
+            toolbarMenuReverseIdMap = rev
+        }
     }
+
+    // endregion
+
+    // region Header lifecycle
+
+    private fun resetHeader() {
+        appBarLayout?.let {
+            detachAppBarListeners(it)
+            removeView(it)
+        }
+        appBarLayout = null
+        toolbarMenuForwardIdMap = emptyMap()
+        toolbarMenuReverseIdMap = emptyMap()
+    }
+
+    private fun removeHeader() {
+        resetHeader()
+        removeContentBehavior()
+        requestLayout()
+    }
+
+    // endregion
+
+    // region Content behavior
+
+    internal fun setContentBehavior() {
+        val params = stackScreenWrapper.layoutParams as LayoutParams
+        if (params.behavior == null) {
+            params.behavior =
+                StackHeaderScrollingViewBehavior { contentTop, _ ->
+                    stackScreen.updateStateIfNeeded(y = contentTop)
+                }
+            stackScreenWrapper.layoutParams = params
+            stackScreenWrapper.requestLayout()
+        }
+    }
+
+    internal fun removeContentBehavior() {
+        val params = stackScreenWrapper.layoutParams as LayoutParams
+        if (params.behavior != null) {
+            params.behavior = null
+            stackScreenWrapper.layoutParams = params
+            stackScreen.updateStateIfNeeded(y = 0)
+            stackScreenWrapper.requestLayout()
+        }
+    }
+
+    // endregion
+
+    // region Shadow state synchronization
+
+    private val appBarOffsetListener =
+        AppBarLayout.OnOffsetChangedListener { _, _ ->
+            syncShadowState()
+        }
+
+    private val appBarLayoutChangeListener =
+        View.OnLayoutChangeListener { _, _, _, _, _, _, _, _, _ ->
+            syncShadowState()
+        }
+
+    private fun attachAppBarListeners(appBar: StackHeaderAppBarLayout) {
+        appBar.addOnOffsetChangedListener(appBarOffsetListener)
+        appBar.addOnLayoutChangeListener(appBarLayoutChangeListener)
+    }
+
+    private fun detachAppBarListeners(appBar: StackHeaderAppBarLayout) {
+        appBar.removeOnOffsetChangedListener(appBarOffsetListener)
+        appBar.removeOnLayoutChangeListener(appBarLayoutChangeListener)
+    }
+
+    private fun syncShadowState() {
+        val delegate = currentDelegate ?: return
+        val provider = currentProvider ?: return
+        val appBar = appBarLayout ?: return
+
+        val configOffset = if (provider.transparent) appBar.top else appBar.top - appBar.bottom
+
+        delegate.updateHeaderFrame(
+            appBar.width,
+            appBar.height,
+            configOffset,
+        )
+
+        updateSubviewOffsets(appBar, provider)
+    }
+
+    private fun updateSubviewOffsets(
+        appBar: StackHeaderAppBarLayout,
+        config: StackHeaderConfigurationProviding,
+    ) {
+        config.leadingSubview?.let { updateSubviewOffset(it, appBar) }
+        config.centerSubview?.let { updateSubviewOffset(it, appBar) }
+        config.trailingSubview?.let { updateSubviewOffset(it, appBar) }
+        config.backgroundSubview?.let { updateSubviewOffset(it, appBar) }
+    }
+
+    private fun updateSubviewOffset(
+        subview: StackHeaderSubviewProviding,
+        appBar: StackHeaderAppBarLayout,
+    ) {
+        val view = subview.view
+        if (view.width == 0 && view.height == 0) return
+
+        val appBarPos = IntArray(2)
+        val subviewPos = IntArray(2)
+        appBar.getLocationInWindow(appBarPos)
+        view.getLocationInWindow(subviewPos)
+
+        subview.updateContentOriginOffset(
+            x = subviewPos[0] - appBarPos[0],
+            y = subviewPos[1] - appBarPos[1],
+        )
+    }
+
+    // endregion
+
+    // region Teardown
 
     internal fun tearDown() {
-        removeCallbacks(headerUpdateRunnable)
-        isHeaderUpdatePending = false
+        resetHeader()
 
         stackScreenWrapper.removeView(stackScreen)
 
-        currentConfig?.removeDelegate(headerConfigDelegate)
-        currentConfig = null
+        currentProvider?.setConfigObserver(null)
+        currentProvider = null
+        currentDelegate = null
 
-        stackScreen.onHeaderConfigAttachListener
+        stackScreen.onHeaderConfigurationAttachListener
             ?.get()
             ?.takeIf { it === onHeaderConfigAttach }
-            ?.let { stackScreen.onHeaderConfigAttachListener = null }
-
-        headerCoordinator.tearDown(this)
+            ?.let {
+                stackScreen.onHeaderConfigurationAttachListener = null
+            }
     }
+
+    // endregion
 }
