--- android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/config/OnHeaderConfigAttachListener.kt
@@ -1,5 +1,0 @@
-package com.swmansion.rnscreens.gamma.stack.header.config
-
-internal fun interface OnHeaderConfigAttachListener {
-    fun onHeaderConfigAttach(config: StackHeaderConfigProviding?)
-}
