+++ android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/config/OnHeaderConfigurationAttachListener.kt
@@ -1,0 +1,8 @@
+package com.swmansion.rnscreens.gamma.stack.header.config
+
+internal fun interface OnHeaderConfigurationAttachListener {
+    fun onHeaderConfigAttach(
+        provider: StackHeaderConfigurationProviding?,
+        delegate: StackHeaderDelegate?,
+    )
+}
