+++ android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/config/StackHeaderConfigurationObserver.kt
@@ -2,8 +2,11 @@ package com.swmansion.rnscreens.gamma.stack.header.config
 
 import com.swmansion.rnscreens.gamma.stack.header.toolbar.StackHeaderToolbarMenuItemOptions
 
-interface StackHeaderConfigDelegate {
-    fun onConfigChange(config: StackHeaderConfigProviding)
+interface StackHeaderConfigurationObserver {
+    fun onConfigChanged(
+        config: StackHeaderConfigurationProviding,
+        flags: StackHeaderUpdateFlags,
+    )
 
     fun onMenuItemUpdate(
         id: String,
