+++ android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/config/StackHeaderConfigurationProviding.kt
@@ -4,7 +4,7 @@ import android.graphics.drawable.Drawable
 import com.swmansion.rnscreens.gamma.stack.header.subview.StackHeaderSubviewProviding
 import com.swmansion.rnscreens.gamma.stack.header.toolbar.StackHeaderToolbarMenuItemConfig
 
-interface StackHeaderConfigProviding {
+interface StackHeaderConfigurationProviding {
     val type: StackHeaderType
     val title: String
     val hidden: Boolean
@@ -24,18 +24,7 @@ interface StackHeaderConfigProviding {
     val trailingSubview: StackHeaderSubviewProviding?
     val backgroundSubview: StackHeaderSubviewProviding?
     val toolbarMenuItems: List<StackHeaderToolbarMenuItemConfig>
-
     val isRTL: Boolean
 
-    fun updateHeaderFrame(
-        width: Int,
-        height: Int,
-        contentOffsetY: Int,
-    )
-
-    fun onMenuItemClick(id: String)
-
-    fun setDelegate(delegate: StackHeaderConfigDelegate)
-
-    fun removeDelegate(delegate: StackHeaderConfigDelegate)
+    fun setConfigObserver(observer: StackHeaderConfigurationObserver?)
 }
