+++ android/src/main/java/com/swmansion/rnscreens/gamma/stack/header/config/StackHeaderDelegate.kt
@@ -1,0 +1,11 @@
+package com.swmansion.rnscreens.gamma.stack.header.config
+
+interface StackHeaderDelegate {
+    fun updateHeaderFrame(
+        width: Int,
+        height: Int,
+        contentOffsetY: Int,
+    )
+
+    fun onMenuItemClick(id: String)
+}
